[Fleet] Improve tests around upgrade
