+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/edit_package_policy_page/components/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { UpgradeStatusCallout } from './upgrade';
