+++ x-pack/plugins/fleet/server/services/package_policy.test.ts
@@ -114,6 +114,36 @@ async function mockedGetPackageInfo(params: any) {
   if (params.pkgName === 'apache') pkg = { version: '1.3.2' };
   if (params.pkgName === 'aws') pkg = { version: '0.3.3' };
   if (params.pkgName === 'endpoint') pkg = {};
+  if (params.pkgName === 'test') {
+    pkg = {
+      version: '1.0.2',
+    };
+  }
+  if (params.pkgName === 'test-conflict') {
+    pkg = {
+      version: '1.0.2',
+      policy_templates: [
+        {
+          name: 'test-conflict',
+          inputs: [
+            {
+              title: 'test',
+              type: 'logs',
+              description: 'test',
+              vars: [
+                {
+                  name: 'test-var-required',
+                  required: true,
+                  type: 'integer',
+                },
+              ],
+            },
+          ],
+        },
+      ],
+    };
+  }
+
   return Promise.resolve(pkg);
 }
 
@@ -1109,6 +1139,10 @@ describe('Package policy service', () => {
 
       expect(result.name).toEqual('test');
     });
+  });
+
+  describe('delete', () => {
+    it('should allow to delete a package policy', async () => {});
   });
 
   describe('runDeleteExternalCallbacks', () => {
@@ -3381,6 +3415,72 @@ describe('Package policy service', () => {
         packagePolicyService.getUpgradePackagePolicyInfo(savedObjectsClient, 'package-policy-id')
       ).rejects.toEqual(new IngestManagerError('Package notinstalled is not installed'));
     });
+  });
+});
+
+describe('getUpgradeDryRunDiff', () => {
+  let savedObjectsClient: jest.Mocked<SavedObjectsClientContract>;
+  beforeEach(() => {
+    savedObjectsClient = savedObjectsClientMock.create();
+    appContextService.start(createAppContextStartContractMock());
+  });
+  afterEach(() => {
+    appContextService.stop();
+  });
+  it('should return no errors if there is no conflict to upgrade', async () => {
+    const res = await packagePolicyService.getUpgradeDryRunDiff(
+      savedObjectsClient,
+      'package-policy-id',
+      {
+        id: '123',
+        name: 'test-123',
+        package: {
+          title: 'test',
+          name: 'test',
+          version: '1.0.1',
+        },
+        namespace: 'default',
+        inputs: [
+          {
+            id: 'toto',
+            enabled: true,
+            streams: [],
+            type: 'logs',
+          },
+        ],
+      } as any,
+      '1.0.2'
+    );
+
+    expect(res.hasErrors).toBeFalsy();
+  });
+
+  it('should no errors if there is a conflict to upgrade', async () => {
+    const res = await packagePolicyService.getUpgradeDryRunDiff(
+      savedObjectsClient,
+      'package-policy-id',
+      {
+        id: '123',
+        name: 'test-123',
+        package: {
+          title: 'test',
+          name: 'test-conflict',
+          version: '1.0.1',
+        },
+        namespace: 'default',
+        inputs: [
+          {
+            id: 'toto',
+            enabled: true,
+            streams: [],
+            type: 'logs',
+          },
+        ],
+      } as any,
+      '1.0.2'
+    );
+
+    expect(res.hasErrors).toBeTruthy();
   });
 });
